pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "unboundapi"
version = "0.2.1"
description = "REST Api for the Unbound DNS Server"
authors = ["Arthur Réauté <arthur@platy.bzh>"]
readme = "README.md"

[tool.poetry.scripts]
unbound-cli = "unboundapi.main:cli"
unbound-api = "unboundapi.server:start"

[tool.poetry.dependencies]
python = ">=3.10,<4"
datetime = ">=5.5"
click = "^8.2.1"
flask = "^3.1.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.2.2"
flask-jwt-extended = "^4.7.1"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
pyright = "^1.1.335"
pytest = "^8.1.1"
tox = "^4.14.2"

[tool.pyright]
venvPath = "."
venv = ".venv"
